requirements.txt
faiss-cpu==1.7.4
Flask==3.0.2
numpy==1.26.1
openai==1.12.0
pandas==2.1.3
sentence-transformers==2.3.1
